API-650: Move VO in Model\VO
