+++ src/Akeneo/Apps/back/Domain/Model/Write/App.php
@@ -5,6 +5,9 @@ declare(strict_types=1);
 namespace Akeneo\Apps\Domain\Model\Write;
 
 use Akeneo\Apps\Domain\Model\ClientId;
+use Akeneo\Apps\Domain\Model\ValueObject\AppCode;
+use Akeneo\Apps\Domain\Model\ValueObject\AppLabel;
+use Akeneo\Apps\Domain\Model\ValueObject\FlowType;
 
 /**
  * @author Romain Monceau <romain@akeneo.com>
