+++ src/Akeneo/Apps/back/Domain/Validation/App/AppCodeMustBeValid.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Apps\Domain\Validation\App;
 
-use Akeneo\Apps\Domain\Model\Write\AppCode;
+use Akeneo\Apps\Domain\Model\ValueObject\AppCode;
 use Symfony\Component\Validator\Context\ExecutionContextInterface;
 
 /**
