+++ src/Akeneo/Apps/back/Domain/Validation/App/AppLabelMustBeValid.php
@@ -4,8 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Apps\Domain\Validation\App;
 
-use Akeneo\Apps\Domain\Model\Write\AppCode;
-use Akeneo\Apps\Domain\Model\Write\AppLabel;
+use Akeneo\Apps\Domain\Model\ValueObject\AppLabel;
 use Symfony\Component\Validator\Context\ExecutionContextInterface;
 
 /**
