+++ src/Akeneo/Apps/back/tests/Unit/spec/Application/Command/CreateAppHandlerSpec.php
@@ -9,7 +9,7 @@ use Akeneo\Apps\Application\Service\CreateClientInterface;
 use Akeneo\Apps\Domain\Exception\ConstraintViolationListException;
 use Akeneo\Apps\Domain\Model\ClientId;
 use Akeneo\Apps\Domain\Model\Write\App;
-use Akeneo\Apps\Domain\Model\Write\FlowType;
+use Akeneo\Apps\Domain\Model\ValueObject\FlowType;
 use Akeneo\Apps\Domain\Persistence\Repository\AppRepository;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
