+++ src/Akeneo/Apps/back/tests/Unit/spec/Application/Query/FetchAppsHandlerSpec.php
@@ -6,7 +6,7 @@ namespace spec\Akeneo\Apps\Application\Query;
 
 use Akeneo\Apps\Application\Query\FetchAppsHandler;
 use Akeneo\Apps\Domain\Model\Read\App;
-use Akeneo\Apps\Domain\Model\Write\FlowType;
+use Akeneo\Apps\Domain\Model\ValueObject\FlowType;
 use Akeneo\Apps\Domain\Persistence\Repository\AppRepository;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
