+++ src/Akeneo/Apps/back/tests/Unit/spec/Domain/Model/ValueObject/AppLabelSpec.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace spec\Akeneo\Apps\Domain\Model\Write;
+namespace spec\Akeneo\Apps\Domain\Model\ValueObject;
 
-use Akeneo\Apps\Domain\Model\Write\AppLabel;
+use Akeneo\Apps\Domain\Model\ValueObject\AppLabel;
 use PhpSpec\ObjectBehavior;
 
 /**
